Rename RepositoryList to RepoList
