+++ feature/repolist/src/main/java/com/treeengineering/repolist/di/RepoListModule.kt
@@ -1,0 +1,12 @@
+package com.treeengineering.repolist.di
+
+import com.treeengineering.repolist.ui.RepoListActionCreator
+import com.treeengineering.repolist.ui.RepoListActionCreatorImpl
+import com.treeengineering.repolist.ui.RepoListStore
+import org.koin.androidx.viewmodel.dsl.viewModel
+import org.koin.dsl.module
+
+val repoListModule = module {
+    factory { RepoListActionCreatorImpl() as RepoListActionCreator }
+    viewModel { RepoListStore() }
+}
