+++ feature/repolist/src/main/java/com/treeengineering/repolist/ui/RepoListAction.kt
@@ -1,0 +1,6 @@
+package com.treeengineering.repolist.ui
+
+import com.treeengineering.core.action.Action
+
+sealed class RepoListAction : Action() {
+}
