+++ feature/repolist/src/main/java/com/treeengineering/repolist/ui/RepoListActionCreator.kt
@@ -1,0 +1,5 @@
+package com.treeengineering.repolist.ui
+
+interface RepoListActionCreator {
+
+}
