+++ feature/repolist/src/main/java/com/treeengineering/repolist/ui/RepoListActionCreatorImpl.kt
@@ -1,0 +1,5 @@
+package com.treeengineering.repolist.ui
+
+class RepoListActionCreatorImpl : RepoListActionCreator {
+
+}
