+++ feature/repolist/src/main/java/com/treeengineering/repolist/ui/RepoListStore.kt
@@ -1,0 +1,7 @@
+package com.treeengineering.repolist.ui
+
+import com.treeengineering.core.store.Store
+
+class RepoListStore : Store() {
+
+}
