--- feature/repositorylist/src/main/java/com/treeengineering/repositorylist/ui/RepositoryListAction.kt
@@ -1,6 +1,0 @@
-package com.treeengineering.repositorylist.ui
-
-import com.treeengineering.core.action.Action
-
-sealed class RepositoryListAction : Action() {
-}
