--- feature/repositorylist/src/main/java/com/treeengineering/repositorylist/ui/RepositoryListActionCreator.kt
@@ -1,5 +1,0 @@
-package com.treeengineering.repositorylist.ui
-
-interface RepositoryListActionCreator {
-    
-}
