--- feature/repositorylist/src/main/java/com/treeengineering/repositorylist/ui/RepositoryListActionCreatorImpl.kt
@@ -1,5 +1,0 @@
-package com.treeengineering.repositorylist.ui
-
-class RepositoryListActionCreatorImpl : RepositoryListActionCreator {
-    
-}
