--- feature/repositorylist/src/main/java/com/treeengineering/repositorylist/ui/RepositoryListStore.kt
@@ -1,7 +1,0 @@
-package com.treeengineering.repositorylist.ui
-
-import com.treeengineering.core.store.Store
-
-class RepositoryListStore : Store() {
-
-}
