+++ settings.gradle
@@ -1,3 +1,3 @@
 include ':app', ':core', ':model'
 include ':data:api', ':data:repository', ':data:db'
-include ':feature:login', ':feature:repositorylist'
+include ':feature:login', ':feature:repolist'
